// src/api/database/class/ProductSTD.ts
import { MediaSTD } from './MediaSTD';
import { IProductPropItem } from '../../../types/portal/server';

export class ProductSTD {
  id: number;
  name: string;
  media: MediaSTD[];
  properties: Record<string, IProductPropItem[]>;
  taxons: Record<string, IProductPropItem>;
  slug: string;
  sku: string;
  sku2: string;
  price: number;
  rf_price: number;
  dprice: number;
  brand_price: string;
  factory_price: string;
  excerpt: string;
  description: string;
  state: string;
  ext_title: string;
  meta_keywords: string;
  meta_description: string;
  deleted_at: string;
  created_at: string;
  updated_at: string;
  units_sold: number;
  last_sale_at: string;
  stock: number;
  rf_manager: number;
  name_en: string;
  name_ru: string;
  bpl: number;
  discount: number;
  promo_pr: number;
  promo_do: string;
  in_stock: number;
  is_new: number;
  is_hit: number;
  is_quanted: number;
  color_text: string;
  smack_text: string;
  aroma_text: string;
  food_text: string;
  postav: string;
  names: string;
  reqcounter: number;
  guid: number;
}